Kj recordable view refactor (#115)
* Convert Workbook.Take::createdTimestamp to modifiedTimestamp Relay

* AssociatedAudio::selectTake() can accept nulls

* WorkbookRepo: Add subscribe() after deleteTake and editTake

* Fix sort in FileNamer

* Add equals() method to Take

In WorkbookRepository, if an AssociatedAudio from the db has a selected take, duplicate takes are being created for the selected take- one ends up in the selected take relay, and the other ends up in the takes relay. Unless this is changed, we need an equals function in Take to compare these different but equivalent takes.

* Add "includeDeleted" parameter to TakeRepo.getByContent

* Move Take subscriptions after the workbook take has been created

Previously, subscriptions to the Take's 'modified' and 'deleted' timestamp were being set up when the workbook take is created in WorkbookRepository. We want to create takes elsewhere in the app as well, so we should set up these subscriptions in the Associated Audio's takesRelay.onNext{} instead. This way, it doesn't matter where the take was created.

* Update unit test in TestWorkbookRepository

* Revert modifiedTimestamp back to createdTimestamp

* Fix duplicate workbook takes being created

The initial selected take needs to use the same workbook take as the rest of the app, otherwise we are setting up subscriptions to the deletedTimestamp that do not get called when the selected take is deleted.

* Revert "Fix duplicate workbook takes being created"

This reverts commit f3a2e5bc28e90affc5c24e5a684cde9281cca25d.

* Fix bug in workbookTake construction

* Revised fix for duplicate workbook takes being created

We don't really need the modelToWbTakeMap, and creating it actually means that the weak references might not go away since the workbook takes are a key in one WeakHashMap and the model takes are a key in another WeakHashMap. We would have to make one of the hash maps contain a WeakReference value for this to work.

* Update .travis.yml

Set distro to Trusty to use oracle jdk 8

* PR Comments