Takes Page Fixes (#475)
* Speed up initial loading of alternate takes

* Fixed drag target bounds

Previously, dragging a take near the drag target could trigger select take even if there was no overlap. Now, there must be overlap to trigger select take. Additionaly, the trigger used to depend on where the mouse is when it is released. Now, it depends on whether the take being dragged overlaps the drag target.

* Temporary fix for project audio directory

* Temporary fix to set active resource slug property

* Using DirectoryProvider.getProjectAudioDirectory instead

* Remove chapter from project audio directory

* Fix deleteProject

* Format CollectionRepository.kt