Update AudioPlugin.kt
Addresses OT-39
The clapper library creates a list of all classes in a jar and then allows filters against it. The subclass filter works by looking at all classes in the found list that have a super class, and then look to see if that super class is in the list of found classes. Then, of those classes with a super class contained in the found classes, it looks for which contain the subclass the user is filtering on.

This by definition can only work if the subclass and the super class both exist in the classpath of the jar; meaning it has to be a fat jar so that the workbook plugin dependency library gets exported as well. Previously this worked because the jar command was executed on the entire project (including both workbook and recording apps) and just changed the main class specified in the manifest. This was unintentional.

As a result, the library is used to pull classes out, but then filtering is done manually rather than by the library.

The URL plugin loader was also needed as the regular class loader only looks to find what is on the main apps classpath (which, again, everything previously had included and exported all code so there wasn't actual separation and the plugin philosophically wasn't really working before)