Allow for smaller width resource take cards #511 (#8)
*  Allow for smaller width resource take cards #511 

"Currently, resource take cards on the right side of the record resource fragment are spilling past the edge of the available space, causing a horizontal scroll bar to appear at the bottom. This PR allows the take card to be smaller than its maximum width. Effectively, on small screens, the selected take area is now smaller than the cards in the alternate takes list on the right side. This PR does ensure that the take being dragged is the same size as the selected take (/drag target) area.

More work needs to be done to ensure that the width of the selected take is the same as the width of the content text. Additonally, more work would need to be done to ensure that the selected take area and the alternate takes are the same size as eachother but still resizable.

Corresponds to OT-23" -Kevin Joslyn

* Merge remote-tracking branch 'origin/dev' into kj-take-card-size