Move gradle dependency, change to WACS
Changes the build order dependencies for the workbook app from the recorder app to the workbook app. I would move the task too, but it would change since the jar variable or command would change contexts, so presumably the whole task code would need to be different.

Changes the location for retrieving the ULB to be from WACS