make sure copy assets task is run, fix test, unused import
set up assets to publish to a maven repo
added a finalizeBy to make sure the copy task is done

Changed numbers in the test due to a discrepancy in the verse counts of en_ulb between the snapshot and what's pulled from wa_catalog. This should be a temporary fix until a real count is established?

Removed an unused import