Bug fixes, minor logging tweaks
Forgot a blocking await to trigger ImportLanguages and ImportUlb

Went back to a foreach blocking await over reduce in initialize app; the reduce happens immediately, so the progress needed to be passed along through the observables; ultimately didn't feel like it added value to continue messing with right now.

Changed the progress check for closing the splash screen and opening the app to be within a doOnComplete block and moved the code into a function.