Jbs default plugins (#11)
* Add ocenaudio default plugin

* Copy Recorder jar into workbookapp resources

Adds a task to copy the generated recorderapp jar to the resources of the workbook app.

* Import Default Recorder

* Download ULB into resources

* Import en_ulb on install from resources

* Move gradle dependency, change to WACS

Changes the build order dependencies for the workbook app from the recorder app to the workbook app. I would move the task too, but it would change since the jar variable or command would change contexts, so presumably the whole task code would need to be different.

Changes the location for retrieving the ULB to be from WACS

* Test importing the downloaded WACS ulb (and import as stream)

Tests importing from a resource stream and the downloaded (live) WACS resource container.

We know that the provided one in the test works, so presumably the live one should match

* Change to WA-Catalog

* Add an assets subproject for shared resources

Creates an assets subproject for common shared resources.

This project currently fetches the english ulb and langnames.json to bundle with the app.

To use this, merge the resources srcDirs within a sourcesets block

* Add initialization table

Adds an initialization table for entering various steps of initialization to help with retrying on error or future migrations

Currently if individual initialization steps fail the app will mark initialized and not re-attempt

* Use initializer repo for app initialization

Intent to split off each major initialization into a domain object, currently this likely is full of race conditions

* Fix errors from copy paste

* TornadoFX deprecation fixes

observableList renamed to observableListOf
values.observable renamed to values.asObservable()

* Rename to match interface name to avoid warning

* make sure copy assets task is run, fix test, unused import

set up assets to publish to a maven repo
added a finalizeBy to make sure the copy task is done

Changed numbers in the test due to a discrepancy in the verse counts of en_ulb between the snapshot and what's pulled from wa_catalog. This should be a temporary fix until a real count is established?

Removed an unused import

* integrationTest depends on copyToResources

also fixed some typos

* test

* Initializers

Sets up initializers for the various default configurations and tasks that need to be set up. This includes importing languages, audio plugins, resource containers, and cleaning the database of any takes not matching the filesystem.

Logging has also been added. Simple implementation is used for now before learning configration necessary for log4j2. Simple will just output to the console.

* fix lint error

* Add end of string to renaming assets regex

Adds the end of string anchor to help ensure this actually hits the end of the filename

* Match exactly the ulb or json file name patterns for renaming

Rather than making assumptions about possible file name patterns, at least for now match what should be expected for the ulb zip and json

* Update base git ignore

Updates the base git ignore to ignore the build directories of all subprojects

* Update the assets gitignore

Removes what should already be ignored from the base git ignore

* Clean up gitignore

Removes the gitignore from assets and moves its rule to the top level, removes redundant rules

* Remove subproject gitignore files

These should be covered now by changes to the rules in the top level gitignore

* Jbs installables (#14)

Installables are a type of initializable task that only needs to be performed once on first run of the application, or when the version of the installable is increased. These carry the information required to be stored to the database, and "install" themselves to the InstalledEntityRepository.

* Remove download plugin from dependencies.gradle

Unused now because of switching to using ivy repositories to download files
(The plugin using this version variable was already removed)

* Better naming of the constant value for the en_ulb filename

* Bug fixes, minor logging tweaks

Forgot a blocking await to trigger ImportLanguages and ImportUlb

Went back to a foreach blocking await over reduce in initialize app; the reduce happens immediately, so the progress needed to be passed along through the observables; ultimately didn't feel like it added value to continue messing with right now.

Changed the progress check for closing the splash screen and opening the app to be within a doOnComplete block and moved the code into a function.

* Merge branch 'dev' into jbs-default-plugins

* Jbs test initializers (#16)

* WIP

* Linter fixes

* Paths as constants, remove doOnError

* Log and open application on initializer error

* Remove unneeded Completable.complete()

* Convert map to blockingsubscribe

* Merge branch 'dev' into jbs-default-plugins