Configure log4j2 to write to a file (#19)
Writes logs to a file.
When logs reach 128KB they will move to an autoincremented file that gets Zipped. Uses the default rollover of deleting the oldest of 4.

The configuration class is written in Java due to the log4j2 API using recursive generics that Kotlin isn't able to type infer.