Audio Buffer Player
Creates an audio player that reads small buffers at a time while playing the audio file. The player uses AudioFileReaders to request buffers of PCM data to play. These readers are responsible for maintaining positition through playback, a mechanism for arbitrary seeking, and (if necessary) decoding the audio into PCM.

This commit provides a WavFileReader implementation for playback of wav files.

TODO: implement seek for WavFileReader and the AudioBufferPlayer.