Filter source collection by project id
Currently, the findSourceCollection method finds all collections that match a particular resource container source. The RC source just contains the RC id, version, and language. As a result, this returns all project collections associated with that resource container, which in the case of the ULB is 66. This then does a first or null, resulting in always returning Genesis.

This fix adds the project as a parameter to the method, so that the projects can then be filtered further by the project slug.