Jbs buffer player (#22)
* Audio Buffer Player

Creates an audio player that reads small buffers at a time while playing the audio file. The player uses AudioFileReaders to request buffers of PCM data to play. These readers are responsible for maintaining positition through playback, a mechanism for arbitrary seeking, and (if necessary) decoding the audio into PCM.

This commit provides a WavFileReader implementation for playback of wav files.

TODO: implement seek for WavFileReader and the AudioBufferPlayer.

* WIP

* Seek

* WIP cue chunk support

* WIP

* Parse and write cue points

* Add test for cue chunk

tests adding cue points and reading them back

* Add more texts, fix null character trim

* Write and read metadata

* test various file lengths

* Merge pull request #21 from Bible-Translation-Tools/jbs-wav-metadata

Jbs wav metadata

* Refactor to merge wavfile changes to common

Isolated the changes to WavFile and WavOutputStream so as to not have to make changes to other places in the app until it was finished, now it needed to be moved back to common.

* More package moving, common wav to common audio wav

* Remove unused imports

* Merge branch 'dev' into jbs-buffer-player

* Move packages

Moves the packages in order to show diffs

* Update kdoc comment

* PR fixes

* PR Fix

* Remove AudioBufferPlayer try/catch and remove lateinit

Removed the try/catch that swallows the exception here, as this should be handled at a higher level.

Removes the need for a lateinit

* Merge branch 'dev' into jbs-buffer-player