Delete container on failed import, not directory
Previously this was recursively deleteing the source directory, for example:
/Orature/rc/src/Door43 World Missions Community/en_ulb/v12

when importing a resumable project, because the resumbable project would itself contain a copy of the ULB which would get ImportResult.ALREADY_EXISTS. This would trigger a delete recursively of the entire directory, and while it would delete the zip attempting to be imported that is a duplicate, it would also delete the existing imported source as well, causing file not found exceptions when opening the resource container in the future.

This fix only deletes the resource container attempting to be imported if it fails.