Merge dev into gridview
* Added Exception Dialog View (#52)

* Added Exception Dialog View

* Refactored: Removed tornadofx dependancy

* Refactored: Removed tornadofx dependancy

* Refactored: Converted to control

* storing FontIcon in variables

* Mm exception handler dialog (#55)

* show custom dialog on exception thrown

* Implemented sending report feature

* Fixed error

* Fixed linter errors

* Put github credentials to github.properties file created during build process

* add crash reporting token to build

* Added tq labels (OT-109) (#57)

* Added tq labels (OT-109)

* Make scene background transparent (#58)

* Remember position of the last selected element and scroll to it on re… (#59)

* Remember position of the last selected element and scroll to it on return

* Made setting selected groupCardItem asynchronous

* Jira issue: OT119 (#56)

* Jira issue: OT119

* bind text wrappingWidth to the tabPane width

* Jbs merge source audio (#49)

* Source audio branch

Moves the source audio branch since it was based on a very out of date dev.

Still in progress

* New Project Card

* Try to add drop shadow

* Swap project cards

* Add project styling and set the cell dimensions in the datagrid

TODO: Find a way to bind the dimensions to the card prefs

* Remove hardcoded css path from fxml, update css

* Add language to Project Card

* Add language to project card

* Initial work on more actions

* Add list of more actions to project card

* Change delete to a use case object

Removes deletion of files from the collection repository, creates a deleteproject usecase to combine the two instead

* linter

* Add actions to project card

* Revert "Change delete to a use case object"

This reverts commit 34e2caa559e00c23bfd45c42ee7020c9393c94bc.

* Revert "Revert "Change delete to a use case object""

This reverts commit de11db81e686f6284c08e3a31ed2b735c583bd23.

* Revert "Merge branch 'jbs-change-cards' into jbs-delete-project"

This reverts commit d818c6ea6315a02587526a5709506e110e8d609a, reversing
changes made to 34e2caa559e00c23bfd45c42ee7020c9393c94bc.

* Remove file deletion from collection repo

* Revert "Revert "Merge branch 'jbs-change-cards' into jbs-delete-project""

This reverts commit b92d2c85b1776391d4a6546273f61569df1931e0.

* Delete resources

Finds resources related to the project and deletes their take entries and files

* Linter fixes

* PR fixes

* Set up parameters to send to plugins

Begin sending additional parameters to plugins: language, book, chapter, chunk info and source audio

* Begin working on reading source audio

* Fix reading source audio

* WIP

* Use inputstream from resource container accessor for getting audio

* WIP

* Remove ownership of Clip from audioplayer node (#51)

* Jbs takes source (#53)

* Takes Source Audio Card

* Source Player Styling

* Source audio playback working

* Cache source audio files

* linter

* Fix button name

* rename unit to chunk

* Fix initial playback after seek

* Pause and play icons properly swapping

* Jbs play verses (#54)

* Add the ability to only load a section of audio from a given file for playback

* remove numbers used for testing

* Adjust duration to account for limited playback section

* Connect section based source audio to plugins

* Load source audio on take management page

* Fix playback on take cards

* Fix audio plugin test

* Consolidate player control logic

Moves the control logic into a class for the skins to use rather than duplicating that functionality

* Remove unused constructor

Co-authored-by: Joe <joe_sarabia@wycliffeassociates.org>

* Update Crowdin configuration file

* Reduce memory usage of exporting a large project (#61)

When exporting, OutOfMemory exceptions were being thrown due to ZipFileSystem:

https://stackoverflow.com/questions/23858706/zipping-a-huge-folder-by-using-a-zipfilesystem-results-in-outofmemoryerror

* Fixes the path to resource audio for exporting (#62)

Resource audio is stored in the source creator directory, and as such export was looking in the wrong folder

* Overwrite existing audio files on import (#63)

When importing, this will overwrite existing files. Otherwise FileAlreadyExistsException is thrown.

* Use the mimetype of the source, not always USFM (#64)

Previously all derived RCs were created as text/usfm, which was incorrect for resources like tq which is text/markdown. As such, importing an exported project would throw a parse error due to using the incorrect parser.

* Delete container on failed import, not directory (#65)

Previously this was recursively deleteing the source directory, for example:
/Orature/rc/src/Door43 World Missions Community/en_ulb/v12

when importing a resumable project, because the resumbable project would itself contain a copy of the ULB which would get ImportResult.ALREADY_EXISTS. This would trigger a delete recursively of the entire directory, and while it would delete the zip attempting to be imported that is a duplicate, it would also delete the existing imported source as well, causing file not found exceptions when opening the resource container in the future.

This fix only deletes the resource container attempting to be imported if it fails.

* Jbs dont export softdeletes (#66)

* Filters out takes that are softdeleted

Since files may be "deleted" from the users perspective but not yet actually deleted from the filesystem, export should filter out files marked for soft-delete

* Linter

* Normalize file separator for comparison

Windows previously wasn't matching deleted takes because the path in the string was opposite that of which is returned by invariantSeparatorsPath which is used to create the lists for selectedChapters and deletedTakes.

* Adds Deleting Project dialog, fixes import dialog (#67)

* Adds Deleting Project dialog, fixes import dialog

Import's progress dialog was not running off the main thread for some reason and as such the onChange listener was not firing until after the import had completed.

Adds a similar dialog for delete.

* Linter

* resource usage typo

* Update Crowdin configuration file

* Restrict export source to only source containers used (#68)

* Copy just the source rc file not the whole directory

Multiple RCs could be imported into the source container directory; we should only grab specifically the RC that is source not everything the directory could possibly have

* Copy File

copyDirectory doesn't work for an individual file as it expects to look at children of the directory to create a relative path. As such, this adds a copyFile method.

* Method Naming Consistency

* Linter

* Merge fixes

* Delete SourceAudioSkin.kt

Co-authored-by: Maxim Murzamuratov <murmax82@gmail.com>
Co-authored-by: Joe <joe_sarabia@wycliffeassociates.org>
Co-authored-by: danparisd <43418833+danparisd@users.noreply.github.com>