Fix on slider track click
As a result of a race condition, updating the slider is a challenge.

A loop will periodically update the slider position based on playback in order to animate the slider. There is no API for getting the position where the slider was clicked, so the slider value has to be queried... which could have been overwritten by the animation rather than the click.

x position of the mouse click is close when scaled by the width, but technically needs to subtract the width of the slider's "thumb"- this width value is private.