Use one recording stream for volume and writing
Use of multiple recorders was causing line exceptions due to race conditions.

This uses only one stream for both the volume bar and writing to a file, in which pausing sends a pause signal to the writer and "active" renderer, but the volume bar continues to get the recoding signal.