Refactor recordables
Make the primary list based on the card model objects, which contains the audio player.

This makes closing and reloading audio easier, and removes the ownership of audio players from the UI controls and moves them back to the viewmodel.

Also uses one list for all items; As a result it is imperative that empty cells are shown for takes that are "selected"