Update isDraggingProperty
it.source is not the scripture take card, but it.gestureSource is, so this is updated.

Also, the drag handlers are not functioning on the card, and onMouseReleased is only caught on MacOS. The event handlers on the fragment are handled properly everywhere, so the dragging property is set there. When the card is dragged outside of that container though (off screen, or to the chrome for example) the DRAG_EXITED handler is executed, effectively making the card visible on the grid afterwards while still dragging. As such the value is set to true in the DRAG_ENTERED handler, because that will fire once the user drags the card back into the fragment region.