Remove duplicate instantiation of AppDatabase
Dagger components were being built in multiple places which resulted in multiple instantiations of the AppDatabase class (and thus, running init multiple times) despite the class intending to be a Singleton object. This was causing multiple calls of database migrations as well as potential race conditions.

Reducing the number of Dagger Component build calls, some reorganizing of code was required, such as providing the directoryProvider to the OratureExceptionHandler class.

Extracted migrations out to a separate DatabaseMigrator class.