Migrates Database to add Marker plugins (#228)
* Add Database Migrations

Versions the database and adds migration code

* Remove duplicate instantiation of AppDatabase

Dagger components were being built in multiple places which resulted in multiple instantiations of the AppDatabase class (and thus, running init multiple times) despite the class intending to be a Singleton object. This was causing multiple calls of database migrations as well as potential race conditions.

Reducing the number of Dagger Component build calls, some reorganizing of code was required, such as providing the directoryProvider to the OratureExceptionHandler class.

Extracted migrations out to a separate DatabaseMigrator class.

* Add a test for database migration

* Version the database to the schema version if set up for the first time

Migrations aren't necessary if the database is being created from the latest schema.