Word align unknown chunk size before seeking over them
Adobe Audition adds a "note" and "ltxt" chunk when placing cue markers. As we don't care about those chunks currently, the appropriate response is to seek over them to get to the next chunk to parse.

This was off by 1 however, as Adobe follows the convention of not including word-alignment padding in the subchunk size.
For example, the size for a "note" chunk was 5, but the data included a 0 padding to make the total data 6. The next "note" chunk was then parsed as the string " not" eventually causing a crash as future chunk sizes were parsed incorrectly due to being offset and not reading from the correct location.

Cue chunks follow a very defined format and thus it is known in advance their total size according to the spec, so those are unchanged.

Parsing the Label chunk could present this issue, and as such it is aligned- however labels in Orature, TranslationRecorder, and BTT Recorder all double-word align strings like this and as such should not present an odd numbered chunk-size.