Load source content in window
Instead of popping out a modal for showing source content while plugins are open, this docks the view in window.

Due to needing to handle content coming back from the plugin, we need to prevent the user from being able to close the window, so we intercept the close request and block it if the plugin is still open.

In order to get the close request handler to work, refactoring needed to occur with respect to the splash screen and primary view.

These changes affected the ProjectTabGroup which was closing the app on back, when the tab was created with tab<ProjectGridFragment>.