Fix scroll based seeking bug
Seeking via scrolling the waveform once audio was played would move the position visually, but playing back would resume where you were before.

This was due to seeking via the audio player, not the audio controller.

Also, moved these to delegate methods in the viewmodel; the view should not be talking to methods on the viewmodel properties to begin with.