Improve adding plugins for Mac (#298)
On MacOS, application binaries are deeper within their .app bundle in the /Applications or /System/Applications directories. These directories are treated as files in the JavaFX FileChooser, and as such cannot be used to select the binary.

We add the ability to edit the text in the path field, rather than rely solely on the result of the FileChooser.

Additionally, we further process the path before setting the value, detect if it is MacOS, and automatically navigate for the user.