Mac Plugins / Fix Marker App Seeking
* Improve adding plugins for Mac (#298)

On MacOS, application binaries are deeper within their .app bundle in the /Applications or /System/Applications directories. These directories are treated as files in the JavaFX FileChooser, and as such cannot be used to select the binary.

We add the ability to edit the text in the path field, rather than rely solely on the result of the FileChooser.

Additionally, we further process the path before setting the value, detect if it is MacOS, and automatically navigate for the user.

* Fix Marker App Seeking (#300)

Seek previous algorithm changed/simplified to fix a bug with seek returning to the beginning of the file if seeking to a previous marker while playing audio.

Pauses/resumes before seeking forward or back to prevent playback interfering with position changes.