seeking/slider tweaks, method renaming
removed the word "absolute" from playback methods. Everything should be relative; if a section of audio is set for playback, then 0 should be the beginning reported, and the end should be with respect to that beginning. Absolute positions within a file or reader should be internal implementation details, not leaked out over public APIs.

Removed padding css from slider controls- Padding affected the slider position, and would cause the thumb to jump forward after a drag or click.

Fixed some bugs related to seeking and positioning based on the previously confusing absolute/relative positioning.