Clamp bounds when opening wav file
Clamps the bounds when opening a wav file so that it does not try to map beyond the actual limits of the file.

This can happen if the audio data length in the wav header is wrong, which could be larger than the file itself.

Also logs the issue, as it is indicative of wav file corruption.

Addresses the stack trace:
java.io.IOException: Channel not open for writing - cannot extend file to required size
	at java.base/sun.nio.ch.FileChannelImpl.map(FileChannelImpl.java:976)
	at org.wycliffeassociates.otter.common.audio.wav.WavFileReader.open(WavFileReader.kt:39)
	at org.wycliffeassociates.otter.jvm.device.audio.AudioBufferPlayer.load(AudioBufferPlayer.kt:59)