Test a build to see what is different...
Orature is importing markdown files in the install4j build. It crashes when launching them from the IDE.

This appears to be because the ZipEntry is null in the OtterZipFile; and this is due to the rootZipFile getEntry method being null when resolving the absolute path. The absolute path variable appears to not have the root however, which is why this fails. Does absolute path contain the root when run from the install4j build? log statements are placed to identify this.