Path.toString not sufficient for absolute path
Path.toString() was used for the purpose of providing an absolute path. While this appears to work with Orature installed through install4j (presumably then the jdk packaged with it) it was not working with intellij using Amazon Corretto.

For example, importing a RC with markdown would get the following path from Path.toString():
content/01.md
While Path.absolutePathString():
en_ulb/content/01.md