Handle negative sizes in chunk parsing
In the case of negative sizes, this error can go undetected by a simple less than size remaining check. However, this can happen if the parser is way out of bounds and arbitrarily interpreting random data as integers.

As such, an additional check is made to verify that the size is above 0.

Modified the try/catch to simply log the error if parsing failed; allowing for the user to still play and or edit the file without crashing the app. Worst case scenario is that metadata is deleted - though it would have been corrupted to begin with.