Clamp bounds when opening wav file (#322)
* Clamp bounds when opening wav file

Clamps the bounds when opening a wav file so that it does not try to map beyond the actual limits of the file.

This can happen if the audio data length in the wav header is wrong, which could be larger than the file itself.

Also logs the issue, as it is indicative of wav file corruption.

Addresses the stack trace:
java.io.IOException: Channel not open for writing - cannot extend file to required size
	at java.base/sun.nio.ch.FileChannelImpl.map(FileChannelImpl.java:976)
	at org.wycliffeassociates.otter.common.audio.wav.WavFileReader.open(WavFileReader.kt:39)
	at org.wycliffeassociates.otter.jvm.device.audio.AudioBufferPlayer.load(AudioBufferPlayer.kt:59)

* Handle negative sizes in chunk parsing

In the case of negative sizes, this error can go undetected by a simple less than size remaining check. However, this can happen if the parser is way out of bounds and arbitrarily interpreting random data as integers.

As such, an additional check is made to verify that the size is above 0.

Modified the try/catch to simply log the error if parsing failed; allowing for the user to still play and or edit the file without crashing the app. Worst case scenario is that metadata is deleted - though it would have been corrupted to begin with.