Markdown Import Fix (#324)
Orature is importing markdown files in the install4j build. It crashes when launching them from the IDE.

This appears to be because the ZipEntry is null in the OtterZipFile; and this is due to the rootZipFile getEntry method being null when resolving the absolute path. The absolute path variable appears to not have the root however, which is why this fails. Does absolute path contain the root when run from the install4j build? log statements are placed to identify this.

* Path.toString not sufficient for absolute path

Path.toString() was used for the purpose of providing an absolute path. While this appears to work with Orature installed through install4j (presumably then the jdk packaged with it) it was not working with intellij using Amazon Corretto.

For example, importing a RC with markdown would get the following path from Path.toString():
content/01.md
While Path.absolutePathString():
en_ulb/content/01.md