Bugfix: playback of marked sections
playback within restricted sections of a file (such as verses in a chapter file) do not need to be adjusted- this is taken care of within the reader. As a result, attempts at offsetting values due to starting position are necessary. For example, playing verse 2 when loaded from a chapter file will begin in the audio player at 0, not the frame position of the chapter file.