Commit code from Max for device switching (#344)
* Commit code from Max for device switching

Code from Max's Device Switching PR, modified slightly to fit better with the audio connection pool.

* Extract AudioFormat to DI Builder

* Update Messages_en.properties

* Change audio output icon

* Bugfix: playback of marked sections

playback within restricted sections of a file (such as verses in a chapter file) do not need to be adjusted- this is taken care of within the reader. As a result, attempts at offsetting values due to starting position are necessary. For example, playing verse 2 when loaded from a chapter file will begin in the audio player at 0, not the frame position of the chapter file.

* Bugfix: glitch on pause/play button

isPlaying is tied to listeners, and actively setting them on button presses causes some conflict. This removes the explicit setting on playback state and instead defers completely to listeners.

* Code Organization

* Pass AudioConnectionFactory to Plugins (#346)

* Pass AudioConnectionFactory to Plugins

* Provide Audio Recorder over Connection Factory (#347)