Fix null pointer exceptions on startup (#385)
Initialization of the SettingsViewModel is happening on app start up (not when the user opens the SettingsView). As a result, the SettingsViewModel attempts to get the list of languages to configure the app locale and language options before the LanguageRepository is configured. This causes valid language slug lookups (such as "en") to return null on the first time the app is launched because parsing of langnames.json and adding entries to the database is not yet done.

As a result, App Locale initialization has been moved to the SplashScreen where other JVM platform specific initialization is happening (specifically, audio devices).

Moved most initialization of the SettingsViewModel to a bind method which is called when the SettingsView is docked.

Also fixed a Maybe function in the AppPreferencesRepository to filter out empty strings, so that when no language is selected initially, there is not an exception thrown.