Move locale initialization to OtterApp
Refactors getting the initial locale to remove a dependency on languageRepository queries which can't happen during the first app launch due to the repository not being initialized with data yet.

Instead, preferredLocale is introduced, which gets a locale from the preferences, if there are none, it checks if the system language is a supported locale, selects and sets it, otherwise it falls back to and sets English.