Close drawers when the overlay is clicked (#401)
* Close drawers when the overlay is clicked

Fires a close drawer event when the overlay is clicked.

Also adds an event listener to deselect all AppBar toggle buttons to make sure none of the buttons stay selected.

* Fix bug that prevents swapping from open drawers

As previously the drawer event was deselecting all buttons, it prevented the behavior of swapping between open drawers.

This only does an all togglegroup deselect if the drawer event came elsewhere.