Namespace verse markers
Writes markers used as verse markers to the label "orature-vm-#"

If none are included when parsing, it will look for lone digits to use as verse markers.

If no lone digits are included, it attempts to parse numbers out of labels as verse markers.

Markers not used for verse markers are saved so that they are preserved on write.