Bugfix: spacing errors in parsed USFM
When scripture text has verses with some text in tags such as poetry, all text blocks are concatenated together. This concatenation can end up combining words if there was not a space before or after the tag.

This fix checks to see if the previous text block ended in whitespace, and if not, it adds a space character.