Bugfix: #405, #406- Improper spacing between words in source text, ignore footnotes/crossreferences (#433)
* Update USFMTools to 1.7.0

* Bugfix: spacing errors in parsed USFM

When scripture text has verses with some text in tags such as poetry, all text blocks are concatenated together. This concatenation can end up combining words if there was not a space before or after the tag.

This fix checks to see if the previous text block ended in whitespace, and if not, it adds a space character.

* Ignore Footnotes and Cross references

Updates USFMTools to allow for ignoring F and X markers (footnotes and cross references).

Changing how spacing works on text block concatenation as spaces often come before and after footnotes, and get treated as part of the text block.