Fix slider resetting to 0 on mouse click
The stop block does not seem necessary since player connections are cleaned up when navigating to a new page.

Having the slider return to 0 on a stop event causes the slider to always reset to 0 when the audio is stopped and the slider is clicked on, rather than keeping the slider where the user seeks to.

Additionally, the default sample rate for the simple audio player is set to DEFAULT (44100), rather than an invalid and unlikely sample rate of 0.