Bugfix: chapter not selecting
The subscription fo the chapter take relay was being initialized in the init block, and the subscription is added to disposables, which is cleared on dock, which means navigating to another page after the initial subscription kills it.

This change moves it to onDock.

Additionally, this is a subscription on the chapter itself, which doesn't need to be triggered any time the card is updated, so it is moved within the active chapter block in onDock.