Fix database migration bugs (#496)
* Add default to license column

Because the license cannot be null, adding the column in the migrator crashes due to trying to add null.

* Remove try/catch for adding columns

Sqlite cannot do addColumnIfNotExists so a try/catch was used instead, however this also intercepts actual exceptions such as constraint failures. As a result the try/catch is removed. This shouldn't be an issue since the schema is versioned so the migrations should not be running on databases that do already have the column.