Bugfix: Duplicating workbooks
Workbooks are retrieved multiple times, resulting in extra subscribers which probably won't get cleaned up.

This change checks to see if a workbook exists in the connection map, and returns it if one is found.

Additionally, equals and hashcode are added to Book, as non-primitives are part of the constructor and thus cannot be a data-class.