Fix Begin/Continue translation
Setting the working chunk is not updated when the user creates a verse by verse translation as it was only done on dock.

Chunking will correctly result in begin translation because of the view re-docking.

While executing this in a map is not ideal, the change to a relay means this no longer completes.