Fix WorkbookRepo Tests and lazy load chunks
As Chunks are now a Relay and not an Observable, some changes had to be made in the WorkbookRepositoryTest to reflect the fact that now blocking get cannot be called and chunks are constructed in a different way.

Additionally, to make them lazily evaluated like they were previously, the function call to set up the relay is wrapped in a lazy block, (and the type is updated in the Chapter constructor parameters as well).