Fix a memory leak related to plugins
Each scope stores a map in tornadofx holding all Views and Fragments for that scope- as such, without deregistering, they will leak every time a plugin is opened.

Also updates tornadofx to a self hosted build which includes the latest changes including a fix to a Fragment related memory leak.