Fix Memory leaks in marker app
Lambdas with view model calls can leak without being held in properties.

Timer moved to the viewmodel instead of the view so that a timer tick does not restore the viewmodel to the components map after the scope is deregistered.

All avenues to close the marker app are now done with firing an event so that the event is properly cleaned up.