WeakListeners for Audio Players (#669)
* Use weak listener for AudioPlayerEvents

Uses a weak listener for AudioPlayerEvents in order to allow for garbage collection of the AudioPlayerController.

* Shared global slider timer with weak listeners

Creates a shared global timer to update sliders; subscribing to this will inform the slider when to update the value.

Subscriptions use weak listeners in order to allow for garbage collection. Previously, tying the slider and audio player to a background thread in each AudioPlayerController via the Observable interval was resulting in a memory and thread leak.

Additionally, updating the isPlaying/isPaused property has been moved to the audio event listener.