Parse chunks that are only headers
If a header has an empty size, previously this would not be included in the list of chunks, producing an Invalid Wav Exception as it would not include an empty data chunk.

This change allows the chunk to parse. Skipping bytes beyond the end of the file will not produce an ioexception, and only the 8 bytes are actually read.