Modify database to support verse bridges
Importing projects using Versification will pre-allocate content for all verses within a chapter.

When verse bridges occur in source text, all but the first content for the bridge should be marked as bridged.

Additionally, in order to know the proper verse range for verse bridges (for use in display), the end verse is now stored in the content entity table.

As end is a reserved keyword in sql, the column name is v_end